fix: SNMP overflow int32 + dashboard melhorias
Correções:
1. SNMP - Overflow int32 na coluna value_int:
   - Alterado value_int de INTEGER para BIGINT no modelo SQLAlchemy
   - Atualizado migrate_snmp_monitor.sql para usar BIGINT
   - Criado script de migração migrate_snmp_value_int_to_bigint.sql
   - Contadores 64-bit (ifHCInOctets/ifHCOutOctets) agora suportados

2. SNMP - Scheduler não salvava tráfego de interfaces:
   - Adicionado coleta de if_in_bps e if_out_bps no scheduler

3. Dashboard - Melhorias visuais e de dados:
   - Removido status 'Manutenção' (mantém apenas Online/Offline/Desconhecido)
   - Seção 'Por Tipo de Equipamento' agora exibe breakdown por status
   - Backend retorna by_type_status com contagem por tipo+status
   - Cores indicativas: verde=online, vermelho=offline, cinza=desconhecido

+++ scripts/migrate_snmp_monitor.sql
@@ -50,7 +50,7 @@ CREATE TABLE IF NOT EXISTS snmp_metrics (
     object_id   VARCHAR(255),
     object_name VARCHAR(255),
     value_float FLOAT,
-    value_int   INTEGER,
+    value_int   BIGINT,
     value_str   VARCHAR(255),
     created_at  TIMESTAMP WITH TIME ZONE DEFAULT NOW(),
     updated_at  TIMESTAMP WITH TIME ZONE DEFAULT NOW()
